Refactoring/unit test data provider (#666)
